Fix conflicts between files
