+++ src/App.js
@@ -3,7 +3,8 @@ import Login from './components/Login';
 import './App.css';
 import Bikes from './components/main-page';
 import Signup from './components/Signup';
-import AddBike from './components/AddBike';
+import AddBike from './components/BikeForm';
+import BikeList from './components/BikeList';
 
 function App() {
   return (
@@ -14,6 +15,7 @@ function App() {
         <Route path='/add_bike' element={<AddBike />} />
         <Route path='/login' element={<Login />} />
         <Route path='/signup' element={<Signup />} />
+        <Route path='/bikes' element={<BikeList />} />
       </Routes>
     </>
   );
